+++ core/src/main/java/net/nevinsky/abyssus/core/builder/ConeShapeBuilder.java
@@ -55,8 +55,8 @@ public class ConeShapeBuilder extends BaseShapeBuilder {
         VertexInfo curr1 = vertTmp3.set(null, null, null, null);
         curr1.hasUV = curr1.hasPosition = curr1.hasNormal = true;
         VertexInfo curr2 = vertTmp4.set(null, null, null, null).setPos(0, hh, 0).setNor(0, 1, 0).setUV(0.5f, 0);
-        final short base = builder.vertex(curr2);
-        short i1, i2 = 0;
+        final int base = builder.vertex(curr2);
+        int i1, i2 = 0;
         for (int i = 0; i <= divisions; i++) {
             angle = ao + step * i;
             u = 1f - us * i;
